create presensi for employee

+++ database/migrations/2024_02_18_110641_create_m_positions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('m_positions', function (Blueprint $table) {
             $table->id();
             $table->string('name', 50);
-            $table->float('salary');
+            $table->decimal('salary', 20, 2);
             $table->string('salary_note', 50)->nullable();
             $table->timestamps();
         });
